feat: comprehensive audit — material loss integration, CPSAT parallel virtualization, LBBD cut pool pruning
Audit scope (90 files, +1740/-1341 lines):

Core model (synaps/model.py):
- Field validators on WorkCenter.speed_factor, max_parallel, Operation.base_duration_min, SetupEntry.setup_minutes, AuxiliaryResource.pool_size, OperationAuxRequirement.quantity_needed
- Predecessor cycle detection (DFS) in ScheduleProblem validator

Dispatch support (synaps/solvers/_dispatch_support.py):
- Added material_loss dict to DispatchContext and SlotCandidate
- Integrated material_loss into find_earliest_feasible_slot and build_dispatch_context

CPSAT solver (synaps/solvers/cpsat_solver.py):
- _virtualize_parallel_work_centers: expands max_parallel>1 WCs into disjunctive lanes for exact SDST
- AUGMECON2 penalty for Pareto-optimality in epsilon-constraint mode
- Symmetry breaking for identical machines (capability_group + speed_factor)
- Greedy warm-start auto-generation with hint injection
- Setup-aware relaxation via min_setup_by_wc

LBBD solver (synaps/solvers/lbbd_solver.py):
- Cut pool pruning: _prune_stale_cuts with max_active_cuts (128) and cut_age_limit (5)
- Setup-cost Benders cuts (actual setup overhead from subproblem)
- Load-balance cuts (Hooker 2007 §7.3)
- Warm-start via prev_assignment_map between iterations
- Gap threshold configurable (default 1%)

Greedy dispatch (synaps/solvers/greedy_dispatch.py):
- K3 material-loss scaling factor in ATCS log-score
- Global material-scale computation
- total_material_loss in ScheduleResult output

Incremental repair (synaps/solvers/incremental_repair.py):
- CPSAT fallback (_cpsat_fallback) for unplaced operations
- Material-loss tie-breaking in candidate selection
- total_material_loss in ScheduleResult output

Registry (synaps/solvers/registry.py):
- CPSAT-EPS-MATERIAL-110 configuration added

Tests:
- tests/test_greedy_dispatch_time_limit.py (2 new tests: T1=TIMEOUT, T2=baseline)
- tests/conftest.py: make_three_state_problem, make_aux_constrained_problem
- tests/test_dispatch_support.py: aux resource blocking, setup time, recompute
- tests/test_router.py: regime routing, latency budget, exact-required, select_solver

CI:
- .pre-commit-config.yaml: ruff v0.11.0, check-yaml, debug-statements
- .gitignore: debug_*/fix_*/inspect_*/apply_*.py, benchmark/instances/

Docs:
- REGRESSION_CLOSURE_2026_05_14.md
- ALNS_INITIAL_SEED_FIX_PLAN_2026_05_01.md updated
- NEXT_WAVE_EXECUTION_PLAN_2026_05_01.md updated
- benchmark/STUDIES_INDEX.md updated

Benchmark scripts updated: study_rhc_50k, study_rhc_500k, study_rhc_alns_doe,
study_rhc_alns_geometry_doe, study_native_acceleration,
study_native_rhc_candidate_acceleration, generate_instances, run_benchmark

+++ benchmark/generate_instances.py
@@ -462,7 +462,7 @@ def _pick_eligible_work_centers(
     work_centers: list[WorkCenter],
     flexibility: float,
 ) -> list[UUID]:
-    base_count = max(1, int(round(len(work_centers) * flexibility)))
+    base_count = max(1, round(len(work_centers) * flexibility))
     jitter = rng.choice([-1, 0, 1]) if len(work_centers) > 2 else 0
     selection_count = min(len(work_centers), max(1, base_count + jitter))
     selected = rng.sample(work_centers, selection_count)
